Add frontend customer orders page
The code is basically a cut & paste of the relevant parts from the gem
solidus_auth_devise.

The rationale for this change is that the orders history page better belongs
inside the solidus main project, because the feature does not have much to do
with the actual authentication process, and because it allows projects that
have an existing authentication system (ie. don't use solidus_auth_devise) to
have the orders page feature out of the box.