Add "merged" state to Spree::Order
When merging two orders together, one of them (the one passed as argument to
`Spree::Order#merge!`) will now transition to the `merged` state.

When an order is in `merged` state, it cannot transition to any other state.

In order for an order to become merged, we need to set the new Active Record
relation `#merged_to_order`. The association allows to keep a reference, for
mostly for debugging purposes, to what order the merged order was merged to.

The order that survives from the merge process has the corresponding inverse
relation `#from_merged_order` that points to the merged order.

One thing to keep in mind is that the order state machine should not define
the `#merge!` method, as it's already defined for the `Spree::Order` class,
see `Spree::Order#merge!(order)`.
This is not actually an issue, as the existing `#merge!` method manages the
state  transition by leveraging the non-persisting method `#merge` that is
implicitly defined by the state machine.

This little inconsistency allows to keep the existing interface, without
needing to deprecate the existing `Spree::Order#merge!` interface and adding
conditional checks during the transitional time on the method arity (the
state machine version of `#merge!` will ignore any passed params).