Update migration task with batch limits
When running the task on large DBs it makes sense to limit the
number of records updated on each transaction in order to avoid
locking the records (or the whole table) for too long.

The default batch size is 100'000 and and it can be overridden
by passing a different value when invoking the task. For example,
this will set the batch size to 1000 rows:

rake solidus:migrations:migrate_default_billing_addresses_to_address_book:up[1000]